# g/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transaction-numbers": {
            "get": {
                "tags": [
                    "Transaction Numbers"
                ],
                "summary": "Récupérer tous les numéros de transaction de l'utilisateur",
                "operationId": "fe2c370aa8a50ad46fb38a0b3a523a52",
                "responses": {
                    "200": {
                        "description": "Liste des numéros de transaction de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-transaction-number": {
            "post": {
                "tags": [
                    "Transaction Numbers"
                ],
                "summary": "Ajouter un nouveau numéro de transaction",
                "operationId": "d61003edacc8ffc3a4251d9980e443ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "description": "Numéro de téléphone",
                                        "type": "string",
                                        "format": "numeric"
                                    },
                                    "operator_mobile_id": {
                                        "description": "ID de l'opérateur mobile",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Numéro de transaction ajouté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de succès",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte, des erreurs de validation sont présentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "description": "Liste des erreurs de validation",
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/generate-code-after-save": {
            "post": {
                "tags": [
                    "Transaction Numbers"
                ],
                "summary": "Générer un code après avoir enregistré le numéro de transaction",
                "operationId": "d85be43b36a6c8444799e50548161388",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "description": "Numéro de téléphone",
                                        "type": "string",
                                        "format": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code généré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code_verify": {
                                            "description": "Code de vérification",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-transaction-number/{id}": {
            "put": {
                "tags": [
                    "Transaction Numbers"
                ],
                "summary": "Mettre à jour le numéro de transaction",
                "operationId": "77be613b6ef0462def343e8ba64acbc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du numéro de transaction",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "description": "Numéro de téléphone",
                                        "type": "string",
                                        "format": "numeric"
                                    },
                                    "operator_mobile_id": {
                                        "description": "ID de l'opérateur mobile",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Numéro de transaction mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de succès",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte, des erreurs de validation sont présentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "description": "Liste des erreurs de validation",
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-transaction-number/{id}": {
            "delete": {
                "tags": [
                    "Transaction Numbers"
                ],
                "summary": "Supprimer le numéro de transaction",
                "operationId": "7d3f3a069ad1b8c62fe3714948ac6d75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du numéro de transaction",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Numéro de transaction supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de succès",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}